Show daily and weekly plots of EE and distance travelled
Added feature for participants to visualize their energy expenditure(kCal) in distance travelled(miles). They can now check their current day profile or trend of last seven days. They are also able to check and change their daily goals for the aforementioned two measures.

Added functionality to clean up the archived file base in the phone to free up the storage system. Keeping last 14 days files, all older files that have been uploaded are deleted now. Even if the date of phone switches to a date in future or past, the phone will always have 14 days of file. This way accidentally deleting files not older than 14 days is avoided.

Fixed a bug that did not update the study end date when the start date was updated. So it was possible for end date to come before start date. When the start date is manually set now, the end date is automatically set to be three months after the start date. One can always go and change the end date explicitly.